+++ src/Services/defineAbility.js
@@ -84,4 +84,14 @@ class Grade {
 
 }
 
-export { defineAbilityFor, Tutor, Eval, Assessment, Standard, Grade }
+class Student {
+  constructor(dict) {
+    for (const key in dict) {
+      if (dict.hasOwnProperty(key)) {
+        this[key] = dict[key];
+      }
+    }
+  }
+}
+
+export { defineAbilityFor, Tutor, Eval, Assessment, Standard, Grade, Student }