+++ functions/main.py
@@ -1,0 +1,5 @@
+from firebase_admin import initialize_app
+
+from fn_impl.email import *
+
+app = initialize_app()