+++ src/Services/email.js
@@ -1,43 +1,71 @@
-import { init, send } from "@emailjs/browser";
 import { getDocs, query, collection, where } from "firebase/firestore";
-import { db } from "./firebase";
+import { db, functions } from "./firebase";
+import { httpsCallable } from "firebase/functions";
 
-init('EEoUW9pUcXKWuGEY1');
+const sendAccessRequest = httpsCallable(functions, 'sendAccessRequestEmail');
+const sendAccessGranted = httpsCallable(functions, 'sendAccessGrantedEmail');
+const sendEvalOwnershipRequest = httpsCallable(functions, 'sendEvalOwnershipRequestEmail');
 
-function sendAuthRequestEmail(userName, userEmail) {
-    let adminEmails = [];
-    getDocs(query(collection(db, 'tutors'), where('clearance', '==', 'admin')))
-        .then((querySnapshot) => {
-            querySnapshot.forEach((doc) => {
-                adminEmails.push(doc.data().email);
-            });
-        })
-        .then(() => {
-            console.log(adminEmails)
-            send('service_tb1cwud', 'template_new_access', {
-                send_to: adminEmails.join(';'),
-                user_name: userName,
-                user_email: userEmail,
-            });
-        })
+async function sendAuthRequestEmail(userName, userEmail) {
+    let adminEmails = (await getDocs(query(collection(db, 'tutors'), where('clearance', '==', 'admin')))).docs;
+    
+    try {
+        let result = await sendAccessRequest({
+            admin_list: adminEmails.map((doc) => doc.data().email).join(";"),
+            user_name: userName,
+            user_email: userEmail,
+            host: window.location.host,
+        });
+        if (result.data.error) {
+            throw new Error(result.data.error);
+        } else {
+            return result;
+        }
+    } catch (error) {
+        throw new Error(error);
+    }
 }
 
-function sendAuthApprovedEmail(userName, userEmail) {
-    let adminEmails = [];
-    getDocs(query(collection(db, 'tutors'), where('clearance', '==', 'admin')))
-        .then((querySnapshot) => {
-            querySnapshot.forEach((doc) => {
-                adminEmails.push(doc.data().email);
-            });
-        })
-        .then(() => {
-            console.log(adminEmails)
-            send('service_tb1cwud', 'template_access_granted', {
-                user_email: userEmail,
-                user_name: userName,
-                admins: adminEmails.join(';'),
-            })
-        })
+async function sendAuthApprovedEmail(userName, userEmail) {
+    let adminEmails = (await getDocs(query(collection(db, 'tutors'), where('clearance', '==', 'admin')))).docs;
+
+    try {
+        let result = await sendAccessGranted({
+            admin_list: adminEmails.map((doc) => doc.data().email).join(";"),
+            user_name: userName,
+            user_email: userEmail,
+            host: window.location.host,
+        });
+        if (result.data.error) {
+            throw new Error(result.data.error);
+        } else {
+            return result;
+        }
+    } catch (error) {
+        throw new Error(error);
+    }
 }
 
-export { sendAuthRequestEmail, sendAuthApprovedEmail };
+async function sendEvalOwnershipRequestEmail(userName, evalUrl, evalTutor, evalStudent, evalDate) {
+    let adminEmails = (await getDocs(query(collection(db, 'tutors'), where('clearance', '==', 'admin')))).docs;
+
+    try {
+        let result = await sendEvalOwnershipRequest({
+            admin_list: adminEmails.map((doc) => doc.data().email).join(";"),
+            user_name: userName,
+            eval_url: evalUrl,
+            eval_tutor: evalTutor,
+            eval_student: evalStudent,
+            eval_date: evalDate,
+        });
+        if (result.data.error) {
+            throw new Error(result.error);
+        } else {
+            return result;
+        }
+    } catch (error) {
+        throw new Error(error);
+    }
+}
+
+export { sendAuthRequestEmail, sendAuthApprovedEmail, sendEvalOwnershipRequestEmail };