+++ src/Services/firebase.js
@@ -6,8 +6,8 @@ import { connectFirestoreEmulator } from "firebase/firestore";
 import { getAuth } from "firebase/auth";
 import { connectAuthEmulator } from "firebase/auth";
 import { connectStorageEmulator, getStorage } from "firebase/storage";
-// import { getFunctions } from "firebase/functions";
-// import { connectFunctionsEmulator } from "firebase/functions";
+import { getFunctions } from "firebase/functions";
+import { connectFunctionsEmulator } from "firebase/functions";
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
 
@@ -22,14 +22,14 @@ const db = getFirestore(app, firebaseConfig.databaseId);
 const auth = getAuth();
 auth.tenantId = firebaseConfig?.tenantId || null;
 const storage = getStorage(app);
-// const functions = getFunctions(app);
+const functions = getFunctions(app);
 
 // eslint-disable-next-line no-restricted-globals
 if (location.hostname === 'localhost') {
   connectFirestoreEmulator(db, 'localhost', 8080);
   connectAuthEmulator(auth, 'http://localhost:9099');
   connectStorageEmulator(storage, 'localhost', 9199);
-  // connectFunctionsEmulator(functions, 'localhost', 5001);
+  connectFunctionsEmulator(functions, 'localhost', 5001);
 }
 
-export { app, analytics, db, auth, storage };
+export { app, analytics, db, auth, storage, functions };